Move types to __init__
